タイトルシーンのカメラ対応 各シーンフェードインだけ追加

+++ RideTheFlow/RideTheFlow/src/scene/TitleScene.h
@@ -67,6 +67,6 @@ private:
 
 	/* •ÊƒXƒNƒŠ[ƒ“ */
 	int		screenHandle;
-	int		screenPos;
+	float	screenPos;
 	float	slideTime;
 };
